made imports node compatible and refactored code
