+++ packages/astro-imagetools/utils/getContainerElement.js
@@ -1,5 +1,5 @@
 // @ts-check
-import getAttributesString from "./getAttributes";
+import getAttributesString from "./getAttributesString.js";
 
 export default function getContainerElement({
   tag,
