+++ packages/astro-imagetools/utils/getImgElement.js
@@ -1,6 +1,6 @@
 // @ts-check
 
-import getAttributesString from "./getAttributes";
+import getAttributesString from "./getAttributesString.js";
 
 export default function getImgElement({
   src,
