+++ packages/astro-imagetools/utils/getLayoutStyles.js
@@ -1,4 +1,5 @@
 // @ts-check
+
 export default function getLayoutStyles({
   layout = null,
   isBackgroundImage = false,
