+++ packages/astro-imagetools/utils/getLinkElement.js
@@ -1,5 +1,5 @@
 // @ts-check
-import getAttributesString from "./getAttributes";
+import getAttributesString from "./getAttributesString.js";
 
 export default function getLinkElement({
   images = [],
